chore: API fetchRequest 폴더 구조 변경
