+++ src/app/(with-header)/challenge/personal/[id]/page.tsx
@@ -2,8 +2,8 @@ import { notFound } from 'next/navigation'
 
 import { dehydrate, HydrationBoundary } from '@tanstack/react-query'
 
+import ChallengePersonalDetails from '@app/(with-header)/challenge/personal/[id]/ChallengePersonalDetails'
 import { getPersonalChallengeDetails } from '@features/challenge/api/get-personal-challenge-details'
-import ChallengePersonalDetails from '@features/challenge/components/challenge/personal/details/ChallengePersonalDetails'
 import { QUERY_OPTIONS } from '@shared/config/tanstack-query/query-defaults'
 import { QUERY_KEYS } from '@shared/config/tanstack-query/query-keys'
 import { getQueryClient } from '@shared/config/tanstack-query/queryClient'
