+++ src/features/challenge/hook/useInfiniteGroupChallengeVerifications.ts
@@ -1,7 +1,7 @@
 import { type InfiniteData, useInfiniteQuery } from '@tanstack/react-query'
 
 import { QUERY_KEYS } from '@shared/config/tanstack-query/query-keys'
-import type { ApiResponse } from '@shared/lib/api/fetcher/fetcher'
+import { ApiResponse } from '@shared/lib/api/fetcher/type'
 
 import {
   getGroupChallengeParticipateList,
