+++ src/features/challenge/hook/useInfiniteGroupParticipations.ts
@@ -1,7 +1,7 @@
 import { type InfiniteData, useInfiniteQuery } from '@tanstack/react-query'
 
 import { QUERY_KEYS } from '@shared/config/tanstack-query/query-keys'
-import { ApiResponse } from '@shared/lib/api/fetcher/fetcher'
+import { ApiResponse } from '@shared/lib/api/fetcher/type'
 
 import type { ChallengeResponse } from '../api/participate/group-participant'
 import {
