+++ src/shared/config/tanstack-query/mutation-defaults.ts
@@ -9,7 +9,7 @@ import { Logout } from '@features/member/api/logout'
 import { readAllAlarms } from '@features/member/api/read-all-alarms'
 import { SignUp } from '@features/member/api/signup'
 import { Unregister } from '@features/member/api/unregister'
-import { ApiResponse, ErrorResponse } from '@shared/lib/api/fetcher/fetcher'
+import { ApiResponse, ErrorResponse } from '@shared/lib/api/fetcher/type'
 
 import { MUTATION_KEYS } from './mutation-keys'
 import { QUERY_KEYS } from './query-keys'
