+++ src/shared/lib/api/fetcher/index.ts
@@ -1,0 +1,2 @@
+export * from './fetcher'
+export * from './serverFetchRequest'
